requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
awsebcli==3.19.2
botocore==1.19.31
cairocffi==1.2.0
CairoSVG==2.5.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.4
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-easy-pdf==0.1.1
djangorestframework==3.12.2
future==0.16.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
pypiwin32==223
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
PyYAML==5.3.1
reportlab==3.5.56
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
tinycss2==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.5
